Moved DocListItem to new components folder
